Extract Rhai engine helpers into reusable library
